feat: rename some classes
