+++ app/src/main/java/by/a_makarevich/androidacademyhw1/adapter/MovieAdapter.kt
@@ -7,6 +7,8 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.RatingBar
 import android.widget.TextView
+import androidx.paging.PagingDataAdapter
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import by.a_makarevich.androidacademyhw1.R
 import by.a_makarevich.androidacademyhw1.data.Movie
@@ -14,37 +16,37 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterInside
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 
-class MovieAdapter(private val onClickListenerDetail: OnClickListenerDetail) :
-    RecyclerView.Adapter<MovieViewHolder>() {
+class MovieAdapterPagination(private val onClickListenerDetailPagination: OnClickListenerDetailPagination) :
+    PagingDataAdapter<Movie, RecyclerView.ViewHolder>(REPO_COMPARATOR) {
 
-    private var data = listOf<Movie>()
+    companion object {
+        private  val REPO_COMPARATOR = object : DiffUtil.ItemCallback<Movie>(){
+            override fun areItemsTheSame(oldItem: Movie, newItem: Movie) =
+                oldItem.id == newItem.id
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MovieViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.movies_list_item, null)
-
-        Log.d("MyLog", "onCreateViewHolder")
-        return MovieViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: MovieViewHolder, position: Int) {
-        Log.d("MyLog", "Adapter_onBindViewHolder$position")
-        holder.bind(data[position])
-        holder.itemView.setOnClickListener {
-            onClickListenerDetail.onItemClick(data[position].id)
+            override fun areContentsTheSame(oldItem: Movie, newItem: Movie) =
+                oldItem == newItem
         }
     }
 
-    override fun getItemCount(): Int = data.size
-
-    fun setData(data: List<Movie>) {
-        this.data = data
-        notifyDataSetChanged()
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.movies_list_item, null)
+        Log.d("MyLog", "onCreateViewHolder")
+        return MovieViewHolderPagination(view)
     }
 
+    override fun onBindViewHolder(holderPagination: RecyclerView.ViewHolder, position: Int) {
+        Log.d("MyLog", "Adapter_onBindViewHolder$position")
+        (holderPagination as MovieViewHolderPagination).bind(getItem(position)!!)
+        holderPagination.itemView.setOnClickListener {
+            onClickListenerDetailPagination.onItemClick(getItem(position)!!.id)
+        }
+    }
 }
 
-class MovieViewHolder(itemView: View) :
+class MovieViewHolderPagination(itemView: View) :
     RecyclerView.ViewHolder(itemView) {
+
     private val title: TextView = itemView.findViewById(R.id.title_item)
     private val imageViewMovie: ImageView = itemView.findViewById(R.id.image_view_movie)
     private val minimumAge: TextView = itemView.findViewById(R.id.minimumAge)
@@ -75,6 +77,6 @@ class MovieViewHolder(itemView: View) :
     }
 }
 
-interface OnClickListenerDetail {
+interface OnClickListenerDetailPagination {
     fun onItemClick(movie_id: Int)
 }
