+++ app/src/main/java/by/a_makarevich/androidacademyhw1/api/MovieDetailsResponse.kt
@@ -9,13 +9,13 @@ data class MovieDetailsResponse(
     @SerialName("id")
     val id: Int,
     @SerialName("title")
-    val title: String,
+    val title: String = "",
     @SerialName("overview")
-    val overview: String,
+    val overview: String = "",
     @SerialName("poster_path")
-    val poster: String,
+    val poster: String = "",
     @SerialName("backdrop_path")
-    val backdrop: String,
+    val backdrop: String = "",
     @SerialName("vote_average")
     val ratings: Float,
     @SerialName("vote_count")
