+++ app/src/main/java/by/a_makarevich/androidacademyhw1/ui/movies/FragmentMoviesList.kt
@@ -4,8 +4,6 @@ import android.content.res.Configuration
 import android.os.Bundle
 import android.util.Log
 import android.view.View
-import android.view.View.INVISIBLE
-import android.view.View.VISIBLE
 import android.widget.ProgressBar
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
@@ -19,7 +17,6 @@ import by.a_makarevich.androidacademyhw1.R
 import by.a_makarevich.androidacademyhw1.adapter.MovieAdapterPagination
 import by.a_makarevich.androidacademyhw1.adapter.MovieListItemDecoration
 import by.a_makarevich.androidacademyhw1.adapter.OnClickListenerDetailPagination
-import by.a_makarevich.androidacademyhw1.utils.StatusResult
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.distinctUntilChanged
@@ -32,18 +29,7 @@ class FragmentMoviesList : Fragment(R.layout.fragment_movies_list),
     private var recyclerView: RecyclerView? = null
     private var progressBar: ProgressBar? = null
 
-    /*private var _binding: FragmentMoviesListBinding? = null
-    private val binding get() = _binding!!*/
-
-    //!!!!  private val viewModel by viewModels<FragmentMoviesListViewModel>()
-
-
-    private val viewModel by viewModels<MoviesListViewModelPagination>()
-
-
-    //private val viewModel: FragmentMoviesListViewModel by viewModels { ViewModelFactory() }
-
-    //!!!! private val movieAdapter = MovieAdapter(this)
+    private val viewModel by viewModels<FragmentMoviesListViewModel>()
 
     private val movieAdapter = MovieAdapterPagination(this)
 
@@ -56,7 +42,6 @@ class FragmentMoviesList : Fragment(R.layout.fragment_movies_list),
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        //  _binding = FragmentMoviesListBinding.bind(view)
         initViews(view)
         setUpMovieAdapter()
         setMovieAdapterStateListener()
@@ -86,16 +71,6 @@ class FragmentMoviesList : Fragment(R.layout.fragment_movies_list),
         progressBar = null
     }
 
-    //!!!!!!
-    /* private fun setUpObservers(){
-         viewModel.movieList.observe(this.viewLifecycleOwner) {
-             updateMovieAdapter(it)
-         }
-         viewModel.status.observe(this.viewLifecycleOwner) {
-             setProgressBar(it)
-         }
-     }*/
-    //!!!!!!!
 
     private fun setUpMovieAdapter() {
         Log.d("MyLog", "FragmentMovieList_setUpMovieAdapter")
@@ -116,13 +91,6 @@ class FragmentMoviesList : Fragment(R.layout.fragment_movies_list),
 
     }
 
-    //!!!!!!
-    /*  private fun updateMovieAdapter(movies: List<Movie>) {
-          Log.d("MyLog", "FragmentMovieList_updateMovieAdapter")
-          movieAdapter.setData(movies)
-      }*/
-    //!!!!!!
-
     private fun numberSpans(layoutManager: GridLayoutManager): Int {
         return if (resources.configuration.orientation == Configuration.ORIENTATION_PORTRAIT) {
             layoutManager.spanCount = 2
@@ -133,13 +101,6 @@ class FragmentMoviesList : Fragment(R.layout.fragment_movies_list),
         }
     }
 
-    private fun setProgressBar(status: StatusResult) {
-        when (status) {
-            StatusResult.Loading -> progressBar?.visibility = VISIBLE
-            StatusResult.Success -> progressBar?.visibility = INVISIBLE
-            StatusResult.Error -> progressBar?.visibility = INVISIBLE
-        }
-    }
 
     override fun onItemClick(movie_id: Int) {
         val bundle = Bundle()
@@ -154,3 +115,11 @@ class FragmentMoviesList : Fragment(R.layout.fragment_movies_list),
     }
 
 }
+/*
+    private fun setProgressBar(status: StatusResult) {
+        when (status) {
+            StatusResult.Loading -> progressBar?.visibility = VISIBLE
+            StatusResult.Success -> progressBar?.visibility = INVISIBLE
+            StatusResult.Error -> progressBar?.visibility = INVISIBLE
+        }
+    }*/
