+++ app/src/main/java/by/a_makarevich/androidacademyhw1/ui/movies/FragmentMoviesListViewModel.kt
@@ -1,44 +1,20 @@
 package by.a_makarevich.androidacademyhw1.ui.movies
 
-import android.util.Log
 import androidx.hilt.lifecycle.ViewModelInject
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.paging.PagingData
+import androidx.paging.cachedIn
 import by.a_makarevich.androidacademyhw1.data.Movie
-import by.a_makarevich.androidacademyhw1.data.MovieDBRepository
-import by.a_makarevich.androidacademyhw1.utils.StatusResult
-import kotlinx.coroutines.launch
-import kotlinx.serialization.ExperimentalSerializationApi
+import by.a_makarevich.androidacademyhw1.data.MoviesListRepository
+import kotlinx.coroutines.flow.Flow
 
-@ExperimentalSerializationApi
-class FragmentMoviesListViewModel @ViewModelInject constructor(
-    private val repository: MovieDBRepository
-) : ViewModel() {
+class FragmentMoviesListViewModel @ViewModelInject constructor(private val repository: MoviesListRepository) :
+    ViewModel() {
 
-    private val _status = MutableLiveData<StatusResult>()
-    val status: LiveData<StatusResult> get() = _status
-
-    private val _movieList = MutableLiveData<List<Movie>>(emptyList())
-    val movieList: LiveData<List<Movie>> get() = _movieList
-
-
-    @ExperimentalSerializationApi
-    fun getMoviesRetrofit() {
-        try {
-            viewModelScope.launch {
-                _status.postValue(StatusResult.Loading)
-                _movieList.postValue(repository.loadMovies())
-                _status.postValue(StatusResult.Success)
-            }
-        } catch (exception: Exception) {
-            _status.postValue(StatusResult.Error)
-            Log.d("MyLog", "Exeption $exception")
-        }
+    fun fetchMovies(): Flow<PagingData<Movie>> {
+        return repository.letMoviesFlow()
+            .cachedIn(viewModelScope)
     }
 
-    init {
-        getMoviesRetrofit()
-    }
 }
